# Optimize operations on root of deeply-nested frame tree