--- a/layout/reftests/scrolling/propagated-overflow-style-2-ref.html
+++ b/layout/reftests/scrolling/propagated-overflow-style-2-ref.html
@@ -1,0 +1,15 @@
+<!DOCTYPE html>
+<html>
+<head>
+  <title>
+    Reference case with the root viewport scrollable, via styles on html node.
+  </title>
+  <style>
+    html {
+      overflow: scroll;
+    }
+  </style>
+</head>
+<body>
+</body>
+</html>