--- a/layout/reftests/scrolling/propagated-overflow-style-2e.html
+++ b/layout/reftests/scrolling/propagated-overflow-style-2e.html
@@ -1,0 +1,15 @@
+<!DOCTYPE html>
+<html>
+<head>
+  <title>
+    Testcase with the root viewport scrollable, via styles on body node.
+  </title>
+  <style>
+    body {
+      overflow: scroll;
+    }
+  </style>
+</head>
+<body>
+</body>
+</html>