__An example - but it is individual:__

---
__testcase 1__
(https://bugzilla.mozilla.org/attachment.cgi?id=8843501)

\- actual results:
  \> 2000 ms

\- expected results:
  \< 100 ms

---
__testcase 2__
(https://bugzilla.mozilla.org/attachment.cgi?id=8843503)

\- actual results:
  \> 3000 ms

\- expected results:
  \< 2000 ms

---
__testcase 3__
(https://bugzilla.mozilla.org/attachment.cgi?id=8843521)

\- actual results:
  \> 3000 ms

\- expected results:
  \< 2000 ms

See:
https://bugzilla.mozilla.org/show_bug.cgi?id=1361268
https://bugzilla.mozilla.org/show_bug.cgi?id=1344398
https://bugzilla.mozilla.org/show_bug.cgi?id=1364180
https://bugzilla.mozilla.org/show_bug.cgi?id=1367568

---

I've created the new build (x32, Windows) and tested.