@wolfbeast 2017-09-22
Probably something this might be worth porting to our Pale Moon repo.