@janekptacijarabaci 2017-09-22
Adjustments would have to be made. I leave it to someone else.