@wolfbeast 2017-09-22
Well, d'uh. of course adjustments would have to be made. I did say "porting", not "applying",'grafting" or other terms that might indicate it would apply as-is.
Can you at least indicate if it would generally be portable with adjustments or not?